refactor(esm): convert babel-eslint-plugin
