+++ eslint/babel-eslint-plugin/test/rules/new-cap.js
@@ -1,4 +1,4 @@
-import rule from "../../lib/rules/new-cap.cjs";
+import rule from "../../lib/rules/new-cap.js";
 import RuleTester from "../../../babel-eslint-shared-fixtures/utils/RuleTester.js";
 
 const ruleTester = new RuleTester();
